requirements_whisper.txt
openai-whisper
flask
werkzeug
gunicorn
numpy
torch
torchvision
torchaudio 